refactor: remove now useless DefaultInformable
